Pupmatch Fixed Error && Firebase Deploy
In my Details Component I had a conditional rendering where I was checking if current users's id is the same as the users profile that is currently displayed. But when logged out there is no current user so there was no checking if there is logged in user. I fixed this. I also deployed my project using Firebase